Change the type of aggregate ids to Long

+++ services/ordering/src/main/java/com/eshop/ordering/api/application/commands/CancelOrderCommand.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.api.application.commands;
 
 import an.awesome.pipelinr.Command;
 
-public record CancelOrderCommand(Integer orderNumber) implements Command<Boolean> {
+public record CancelOrderCommand(Long orderNumber) implements Command<Boolean> {
 }
