+++ services/ordering/src/main/java/com/eshop/ordering/domain/events/OrderStatusChangedToStockConfirmedDomainEvent.java
@@ -2,5 +2,5 @@ package com.eshop.ordering.domain.events;
 
 import com.eshop.ordering.domain.seedwork.DomainEvent;
 
-public record OrderStatusChangedToStockConfirmedDomainEvent(Integer orderId) implements DomainEvent {
+public record OrderStatusChangedToStockConfirmedDomainEvent(Long orderId) implements DomainEvent {
 }
